S0C-3A-1A: Add dual run stage 1 scenario and related changes
- Registered new scenario `shadow_verify_dual_run_stage1` for dual run verification.
- Updated `load_builtin_scenarios` to include the new scenario.
- Enhanced documentation with evidence and status updates for dual run tests.
- Introduced a new SQL script for seeding the search index.
- Implemented HTTP request handling and JSON parsing for Elasticsearch interactions.
- Added observability features including tracing for the dual run process.
- Ensured compatibility with existing database and Elasticsearch configurations.
