+++ backend/scripts/cli.py
@@ -1404,415 +1404,112 @@ def _cmd_labs_shadow_verify_dual_run_stage1(args: argparse.Namespace) -> int:
         print("[labs shadow-verify-dual-run-stage1] --strategy must be one of: soft, strict")
         return 2
 
-    scope = "all" if library_id is None else f"library:{library_id}"
-
     es_url = (args.es_url or env.get("ELASTIC_URL") or "http://127.0.0.1:19200").strip().rstrip("/")
     token_default = "dualrun" + re.sub(r"[^0-9A-Za-z]+", "", run_id)
     token = (args.token or token_default).strip() or token_default
 
     def _sanitize_index_name(name: str) -> str:
-        # ES index must be lowercase and avoid special chars.
         safe = re.sub(r"[^a-z0-9_\-]+", "-", name.lower()).strip("-_")
         safe = re.sub(r"-+", "-", safe)
         if not safe:
             safe = "wordloom-search-index"
         return safe[:80]
 
-    es_index = (args.es_index or env.get("ELASTIC_INDEX") or _sanitize_index_name(f"wordloom-search-index-dualrun-{token}")).strip()
+    es_index = (
+        args.es_index
+        or env.get("ELASTIC_INDEX")
+        or _sanitize_index_name(f"wordloom-search-index-dualrun-{token}")
+    ).strip()
     es_index = _sanitize_index_name(es_index)
     recreate_index = bool(args.recreate_index)
 
-    seed_text_prefix = f"{token} "
-    seed_entity_type = "block"  # Postgres FTS provider is block-only.
+    _wg_registry.load_builtin_scenarios()
+    handler = _wg_registry.get(SCENARIO_SHADOW_VERIFY_DUAL_RUN_STAGE1)
 
-    where_parts: list[str] = ["entity_type = :entity_type", "text ILIKE :pattern"]
-    base_params: dict[str, object] = {
-        "entity_type": seed_entity_type,
-        "pattern": f"%{token}%",
-    }
-    if library_id is not None:
-        where_parts.append("library_id = :library_id")
-        base_params["library_id"] = library_id
-
-    where_sql = " AND ".join(where_parts)
-
-    engine = create_engine(database_url)
-    with engine.connect() as conn:
-        # Seed specifically for this drill's query (block-only + token match).
-        existing_for_token = int(
-            conn.execute(text(f"SELECT COUNT(*) FROM search_index WHERE {where_sql}"), base_params).scalar() or 0
-        )
-        need = int(ensure_min_rows) - int(existing_for_token)
-        inserted_rows = 0
-        if need > 0:
-            now = datetime.now(timezone.utc)
-            max_ev = int(conn.execute(text("SELECT COALESCE(MAX(event_version), 0) FROM search_index")).scalar() or 0)
-            rows = []
-            for i in range(int(need)):
-                rows.append(
-                    {
-                        "id": str(uuid.uuid4()),
-                        "entity_type": seed_entity_type,
-                        "library_id": library_id,
-                        "entity_id": str(uuid.uuid4()),
-                        "text": f"{seed_text_prefix}{i}",
-                        "snippet": None,
-                        "rank_score": 0.0,
-                        "created_at": now,
-                        "updated_at": now,
-                        "event_version": int(max_ev + i + 1),
-                    }
-                )
-
-            conn.execute(
-                text(
-                    """
-                    INSERT INTO search_index
-                      (id, entity_type, library_id, entity_id, text, snippet, rank_score, created_at, updated_at, event_version)
-                    VALUES
-                      (:id, :entity_type, :library_id, :entity_id, :text, :snippet, :rank_score, :created_at, :updated_at, :event_version)
-                    """
-                ),
-                rows,
-            )
-            conn.commit()
-            inserted_rows = int(need)
-
-        pg_sql = f"""
-            SELECT entity_id::text AS entity_id,
-                   COALESCE(event_version, 0) AS event_version,
-                   library_id::text AS library_id
-            FROM search_index
-            WHERE {where_sql}
-            ORDER BY COALESCE(event_version, 0) ASC, entity_id::text ASC
-            LIMIT :limit
-        """
-        pg_params = dict(base_params)
-        pg_params["limit"] = candidate_limit
-        pg_rows = conn.execute(text(pg_sql), pg_params).all()
-        pg_candidates = [
-            {
-                "entity_id": str(r[0]),
-                "event_version": int(r[1] or 0),
-                "library_id": (str(r[2]) if (len(r) > 2 and r[2] is not None) else None),
-            }
-            for r in pg_rows
-        ]
-
-    # Strong mutual evidence: emit a machine-searchable stdout probe.
-    probe: dict[str, object] = {
-        "event": "labs.dual_run.stage1.probe",
-        "lab_id": LAB_ID_S2B_2A_2A,
-        "scenario": SCENARIO_SHADOW_VERIFY_DUAL_RUN_STAGE1,
-        "run_id": run_id,
-        "scope": scope,
-        "library_id": library_id,
-        "token": token,
-        "pg_candidates_total": len(pg_candidates),
-        "pg_first": (pg_candidates[0] if pg_candidates else None),
-        "es_url": es_url,
-        "es_index": es_index,
-    }
-    print(json.dumps(probe, ensure_ascii=False, separators=(",", ":")))
-
-    # ES health (best-effort)
-    es_health_status, es_health_payload = _http_json("GET", f"{es_url}", body=None, timeout_s=5.0)
-    es_health_ok = bool(es_health_status == 200)
-
-    # Backfill ES from Postgres via legacy script.
-    backfill_script = LEGACY_SCRIPTS_DIR / "backfill_elastic_search_index.py"
-    backfill_cmd = [
-        _python_exe(),
-        str(backfill_script),
-        "--batch-size",
-        str(backfill_batch_size),
-    ]
-    if recreate_index:
-        backfill_cmd.append("--recreate")
-
-    child_env = env.copy()
-    child_env["DATABASE_URL"] = database_url
-    child_env["ELASTIC_URL"] = es_url
-    child_env["ELASTIC_INDEX"] = es_index
-
-    backfill_proc = subprocess.run(
-        backfill_cmd,
-        cwd=str(REPO_ROOT),
-        env=child_env,
-        capture_output=True,
-        text=True,
-        check=False,
-    )
-
-    backfill_log_path = outdir / "backfill.log"
-    backfill_log_path.write_text(
-        (backfill_proc.stdout or "") + "\n--- stderr ---\n" + (backfill_proc.stderr or "") + "\n",
-        encoding="utf-8",
-    )
-    backfill_exit_code = int(backfill_proc.returncode)
-    backfill_ok = bool(backfill_exit_code == 0)
-
-    def _try_parse_json(payload: str) -> dict[str, object]:
-        try:
-            obj = json.loads(payload) if payload else {}
-            return obj if isinstance(obj, dict) else {"_": obj}
-        except Exception:
-            return {"raw": payload}
-
-    es_query_filters: list[dict[str, object]] = [{"term": {"entity_type": seed_entity_type}}]
-    if library_id is not None:
-        es_query_filters.append({"term": {"library_id": library_id}})
-
-    es_search_body: dict[str, object] = {
-        "query": {
-            "bool": {
-                "must": [{"match": {"text": token}}],
-                "filter": es_query_filters,
-            }
-        },
-        "sort": [{"event_version": "asc"}, {"entity_id": "asc"}],
-        "_source": ["entity_id", "event_version", "entity_type", "library_id"],
-        "size": candidate_limit,
-    }
-
-    es_search_status, es_search_payload = _http_json(
-        "POST",
-        f"{es_url}/{es_index}/_search",
-        body=es_search_body,
-        timeout_s=10.0,
-    )
-    es_search_ok = bool(es_search_status < 400)
-    es_search_obj = _try_parse_json(es_search_payload)
-    hits = (((es_search_obj.get("hits") or {}) if isinstance(es_search_obj, dict) else {}).get("hits") or [])
-    if not isinstance(hits, list):
-        hits = []
-
-    es_candidates: list[dict[str, object]] = []
-    for h in hits:
-        if not isinstance(h, dict):
-            continue
-        src = h.get("_source")
-        if not isinstance(src, dict):
-            continue
-        entity_id = src.get("entity_id")
-        event_version = src.get("event_version")
-        if entity_id is None:
-            continue
-        try:
-            ev = int(event_version or 0)
-        except Exception:
-            ev = 0
-        es_candidates.append({"entity_id": str(entity_id), "event_version": ev})
-
-    es_count_status, es_count_payload = _http_json(
-        "POST",
-        f"{es_url}/{es_index}/_count",
-        body={"query": es_search_body.get("query")},
-        timeout_s=10.0,
-    )
-    es_count_obj = _try_parse_json(es_count_payload)
-    es_count = None
-    if isinstance(es_count_obj, dict) and isinstance(es_count_obj.get("count"), int):
-        es_count = int(es_count_obj["count"])
-
-    pg_ids = [str(c["entity_id"]) for c in pg_candidates]
-    es_ids = [str(c["entity_id"]) for c in es_candidates]
-    if strategy == "strict":
-        parity_ok = bool(pg_ids == es_ids)
-    else:
-        parity_ok = bool(set(pg_ids) & set(es_ids))
-
-    ok = bool(
-        len(pg_candidates) > 0
-        and es_health_ok
-        and backfill_ok
-        and es_search_ok
-        and parity_ok
-    )
-
-    traces_written = False
-    traces_error: str | None = None
-    trace_id: str | None = None
-    span_id: str | None = None
-    span_name = "labs.shadow_verify_dual_run_stage1.probe"
-
-    try:
-        from opentelemetry import trace  # type: ignore
-        from opentelemetry.sdk.resources import Resource  # type: ignore
-        from opentelemetry.sdk.trace import TracerProvider  # type: ignore
-        from opentelemetry.sdk.trace.export import (  # type: ignore
-            SimpleSpanProcessor,
-            SpanExporter,
-            SpanExportResult,
-        )
-        from opentelemetry.sdk.trace.sampling import ALWAYS_ON  # type: ignore
-
-        exported_spans: list[dict[str, object]] = []
-
-        class _JsonSpanExporter(SpanExporter):
-            def export(self, spans) -> SpanExportResult:  # type: ignore[override]
-                for s in spans:
-                    ctx = getattr(s, "context", None)
-                    if ctx is not None:
-                        t_id = f"{int(ctx.trace_id):032x}"
-                        s_id = f"{int(ctx.span_id):016x}"
-                    else:
-                        t_id = ""
-                        s_id = ""
-
-                    attrs = dict(getattr(s, "attributes", {}) or {})
-                    exported_spans.append(
-                        {
-                            "name": getattr(s, "name", ""),
-                            "trace_id": t_id,
-                            "span_id": s_id,
-                            "start_time_unix_nano": int(getattr(s, "start_time", 0) or 0),
-                            "end_time_unix_nano": int(getattr(s, "end_time", 0) or 0),
-                            "attributes": attrs,
-                        }
-                    )
-                return SpanExportResult.SUCCESS
-
-            def shutdown(self) -> None:  # type: ignore[override]
-                return None
-
-        provider = TracerProvider(resource=Resource.create({"service.name": "wordloom-labs"}), sampler=ALWAYS_ON)
-        provider.add_span_processor(SimpleSpanProcessor(_JsonSpanExporter()))
-        trace.set_tracer_provider(provider)
-
-        tracer = trace.get_tracer("wordloom.labs")
-        with tracer.start_as_current_span(span_name) as span:
-            span.set_attribute("event", "labs.dual_run.stage1.probe")
-            span.set_attribute("lab_id", LAB_ID_S2B_2A_2A)
-            span.set_attribute("scenario", SCENARIO_SHADOW_VERIFY_DUAL_RUN_STAGE1)
-            span.set_attribute("run_id", run_id)
-            span.set_attribute("scope", scope)
-            span.set_attribute("library_id", library_id or "")
-            span.set_attribute("token", token)
-            span.set_attribute("pg_candidates_total", len(pg_candidates))
-            span.set_attribute("es_candidates_total", len(es_candidates))
-            span.set_attribute("strategy", strategy)
-            span.set_attribute("ok", bool(ok))
-
-            ctx = span.get_span_context()
-            trace_id = f"{int(ctx.trace_id):032x}"
-            span_id = f"{int(ctx.span_id):016x}"
-
-        provider.force_flush()
-
-        (outdir / "traces.json").write_text(
-            json.dumps(
-                {
-                    "service": "wordloom-labs",
-                    "scenario": SCENARIO_SHADOW_VERIFY_DUAL_RUN_STAGE1,
-                    "run_id": run_id,
-                    "spans": exported_spans,
-                },
-                indent=2,
-                ensure_ascii=False,
-            )
-            + "\n",
-            encoding="utf-8",
-        )
-        traces_written = True
-    except Exception as e:
-        traces_error = f"{type(e).__name__}: {e}"
-        try:
-            (outdir / "traces.json").write_text(
-                json.dumps(
-                    {
-                        "service": "wordloom-labs",
-                        "scenario": SCENARIO_SHADOW_VERIFY_DUAL_RUN_STAGE1,
-                        "run_id": run_id,
-                        "spans": [],
-                        "error": traces_error,
-                    },
-                    indent=2,
-                    ensure_ascii=False,
-                )
-                + "\n",
-                encoding="utf-8",
-            )
-        except Exception:
-            pass
-
-    def _rel_repo(path: Path) -> str:
-        try:
-            return str(path.resolve().relative_to(REPO_ROOT).as_posix())
-        except Exception:
-            return str(path.as_posix())
-
-    result: dict[str, object] = {
-        "lab_id": LAB_ID_S2B_2A_2A,
-        "scenario": SCENARIO_SHADOW_VERIFY_DUAL_RUN_STAGE1,
-        "run_id": run_id,
-        "created_at": time.strftime("%Y-%m-%d %H:%M:%S"),
-        "scope": scope,
-        "inputs": {
-            "token": token,
+    input_payload = dict(vars(args))
+    input_payload.pop("func", None)
+    input_payload.update(
+        {
+            "scenario": SCENARIO_SHADOW_VERIFY_DUAL_RUN_STAGE1,
+            "scope_id": LAB_ID_S2B_2A_2A,
+            "run_id": run_id,
+            "outdir": str(outdir),
+            "env": env,
+            "database_url": database_url,
+            "library_id": library_id,
             "ensure_min_rows": ensure_min_rows,
-            "seed_entity_type": seed_entity_type,
             "candidate_limit": candidate_limit,
+            "backfill_batch_size": backfill_batch_size,
             "strategy": strategy,
             "es_url": es_url,
+            "token": token,
             "es_index": es_index,
             "recreate_index": recreate_index,
-            "backfill_batch_size": backfill_batch_size,
-        },
-        "seed_rows_inserted": int(inserted_rows),
-        "postgres": {
-            "query_sql": pg_sql.strip(),
-            "candidates": pg_candidates,
-        },
-        "elasticsearch": {
-            "health": {"status": int(es_health_status), "ok": bool(es_health_ok), "payload": es_health_payload},
-            "backfill": {
-                "exit_code": int(backfill_exit_code),
-                "ok": bool(backfill_ok),
-                "log_path": _rel_repo(backfill_log_path),
-                "script": _rel_repo(backfill_script),
-            },
-            "search": {
-                "status": int(es_search_status),
-                "ok": bool(es_search_ok),
-                "request": es_search_body,
-                "response_excerpt": {
-                    "hits_total": ((es_search_obj.get("hits") or {}).get("total") if isinstance(es_search_obj, dict) else None),
-                },
-                "candidates": es_candidates,
-            },
-            "count": {"status": int(es_count_status), "count": es_count, "payload": es_count_obj},
-        },
-        "compare": {
-            "pg_ids": pg_ids,
-            "es_ids": es_ids,
-            "parity_ok": bool(parity_ok),
-        },
-        "observability": {
-            "log_probe_emitted": True,
-            "trace_probe": {
-                "span_name": span_name,
-                "trace_id": trace_id,
-                "span_id": span_id,
-                "traces_json_written": traces_written,
-                "error": traces_error,
-            },
-        },
-        "ok": bool(ok),
-    }
+        }
+    )
+    inputs = DrillInputs.model_validate(input_payload)
+    drill = handler(inputs)
+    result = drill.meta or {}
+    write_json(outdir / "_result.json", result)
 
-    (outdir / "_result.json").write_text(json.dumps(result, indent=2, ensure_ascii=False) + "\n", encoding="utf-8")
+    ok = bool(result.get("ok"))
+    scope = str(result.get("scope") or "")
+
+    inputs_obj = result.get("inputs") if isinstance(result, dict) else None
+    if not isinstance(inputs_obj, dict):
+        inputs_obj = {}
+
+    token = str(inputs_obj.get("token") or token)
+    strategy = str(inputs_obj.get("strategy") or strategy)
+
+    seed_rows_inserted = int(result.get("seed_rows_inserted") or 0)
+
+    pg_candidates: list[object] = []
+    postgres_obj = result.get("postgres")
+    if isinstance(postgres_obj, dict):
+        cands = postgres_obj.get("candidates")
+        if isinstance(cands, list):
+            pg_candidates = cands
+    pg_candidates_total = int(len(pg_candidates))
+
+    es_obj = result.get("elasticsearch")
+    es_health_ok = False
+    backfill_ok = False
+    backfill_exit_code = 0
+    es_search_ok = False
+    es_search_status = 0
+    es_candidates_total = 0
+    if isinstance(es_obj, dict):
+        health = es_obj.get("health")
+        if isinstance(health, dict):
+            es_health_ok = bool(health.get("ok"))
+        backfill = es_obj.get("backfill")
+        if isinstance(backfill, dict):
+            backfill_ok = bool(backfill.get("ok"))
+            backfill_exit_code = int(backfill.get("exit_code") or 0)
+        search = es_obj.get("search")
+        if isinstance(search, dict):
+            es_search_ok = bool(search.get("ok"))
+            es_search_status = int(search.get("status") or 0)
+            es_cands = search.get("candidates")
+            if isinstance(es_cands, list):
+                es_candidates_total = int(len(es_cands))
+
+    parity_ok = False
+    compare_obj = result.get("compare")
+    if isinstance(compare_obj, dict):
+        parity_ok = bool(compare_obj.get("parity_ok"))
 
     print("labs-018.shadow_verify_dual_run_stage1")
     print(f"scope={scope}")
     print(f"token={token}")
     print(f"ensure_min_rows={ensure_min_rows}")
-    print(f"seed_rows_inserted={inserted_rows}")
-    print(f"pg_candidates_total={len(pg_candidates)}")
+    print(f"seed_rows_inserted={seed_rows_inserted}")
+    print(f"pg_candidates_total={pg_candidates_total}")
     print(f"es_health_ok={es_health_ok}")
     print(f"backfill_ok={backfill_ok} (rc={backfill_exit_code})")
     print(f"es_search_ok={es_search_ok} (status={es_search_status})")
-    print(f"es_candidates_total={len(es_candidates)}")
+    print(f"es_candidates_total={es_candidates_total}")
     print(f"parity_ok={parity_ok} (strategy={strategy})")
     print(f"ok={ok}")
     print(f"outputs: {outdir}")
